Addressed comments by copilot
